Update obstacle_class.py
I am Naman Sharma and I have done three modifications. Firstly, I have added the comment which summarises the complete code and then I have added a function multiply allows the user to specify the number of decimal places they want in the result. It gives more control over the precision of the result, which may be useful in certain situations lastly at the end I have modified the code to use a try/except block to catch a ValueError if the user enters an input that cannot be converted to an integer. If a ValueError is raised, the code will print an error message and prompt the user to enter a valid integer input. This adds a layer of robustness to the program, making it less likely to crash due to invalid inputs.